made better file structure, fixed standard json
