+++ Standard.json
@@ -1,5 +1,4 @@
 {
-  "name": "Standard",
   "Musical Genres": [
     "Rock",
     "Pop",
